requirements.txt
# Required for functionality
canvasapi>=3.3.0
PyGithub>=2.4.0
gspread>=6.1.2
PyYAML6.0.2

# Optional, for linting
black>=24.8.0
mypy>=1.11.2
pydoclint>=0.5.7
flake8>=7.1.1
flake8-pydocstyle>=0.2.4
flake8-annotations>=3.1.1
types-pyyaml>=6.0.12.20240808